Seperate out BPF Loader helpers
